+++ include/gendil/Meshes/MeshDataStructures/UnstructuredMesh/hexmesh.hpp
@@ -7,7 +7,7 @@
 #include "gendil/Utilities/types.hpp"
 #include "gendil/Meshes/Cells/ReferenceCells/hexcell.hpp"
 #include "gendil/Meshes/Geometries/hypercube.hpp"
-#include "gendil/Meshes/Connectivities/unstructuredconformingconnectivity.hpp"
+#include "gendil/Meshes/MeshDataStructures/UnstructuredMesh/unstructuredconformingconnectivity.hpp"
 #include "gendil/Utilities/View/Layouts/stridedlayout.hpp"
 
 namespace gendil {
@@ -44,9 +44,9 @@ struct HexMesh
 
    template < Integer FaceIndex >
    GENDIL_HOST_DEVICE
-   auto GetFaceNeighborInfo( GlobalIndex cell_index, std::integral_constant< Integer, FaceIndex > face_index ) const
+   auto GetLocalFaceInfo( GlobalIndex cell_index, std::integral_constant< Integer, FaceIndex > face_index ) const
    {
-      return connectivity( cell_index, face_index );
+      return connectivity.GetLocalFaceInfo( cell_index, face_index );
    }
 };