+++ include/gendil/Meshes/MeshDataStructures/UnstructuredMesh/unstructuredmesh.hpp
@@ -1,0 +1,17 @@
+// Copyright GenDiL Project Developers. See COPYRIGHT file for details.
+//
+// SPDX-License-Identifier: (BSD-3-Clause)
+
+#pragma once
+
+// A 1d unstructured mesh.
+#include "gendil/Meshes/MeshDataStructures/UnstructuredMesh/linemesh.hpp"
+
+// A 2d unstructured quad mesh.
+#include "gendil/Meshes/MeshDataStructures/UnstructuredMesh/quadmesh.hpp"
+
+// A 3d unstructured hex mesh.
+#include "gendil/Meshes/MeshDataStructures/UnstructuredMesh/hexmesh.hpp"
+
+// Local face connectivity for conforming unstructured meshes.
+#include "gendil/Meshes/MeshDataStructures/UnstructuredMesh/unstructuredconformingconnectivity.hpp"