+++ include/gendil/Meshes/cell.hpp
@@ -5,6 +5,7 @@
 #pragma once
 
 #include "gendil/Utilities/tensorindex.hpp"
+#include "gendil/Meshes/Connectivities/faceconnectivity.hpp"
 
 namespace gendil {
 
@@ -14,6 +15,7 @@ template <
    typename Cell,
    typename QuadData
 >
+GENDIL_HOST_DEVICE
 void GetValuesAndJacobian(
    const Cell & cell,
    TensorIndex< Cell::Dim > const & quad_index,
@@ -22,6 +24,24 @@ void GetValuesAndJacobian(
    typename Cell::jacobian & J_mesh )
 {
    cell.GetValuesAndJacobian( quad_index, quad_data, X, J_mesh );
+}
+template <
+   typename Cell,
+   CellFaceView Face,
+   typename FaceQuadData
+>
+GENDIL_HOST_DEVICE
+void ComputePhysicalCoordinatesAndJacobian(
+   const Cell & cell,
+   const Face & face,
+   TensorIndex< Cell::Dim > const & quad_index,
+   FaceQuadData const & face_quad_data,
+   typename Cell::physical_coordinates & X,
+   typename Cell::jacobian & J_mesh )
+{
+   constexpr Integer local_face_index = Face::local_face_index_type::value;
+   const auto & local_face_quad_data = std::get< local_face_index >( face_quad_data );
+   cell.GetValuesAndJacobian( quad_index, local_face_quad_data, X, J_mesh );
 }
 
 }